+++ packages/python/plotly/_plotly_utils/basevalidators.py
@@ -53,7 +53,7 @@ def to_scalar_or_list(v):
         return v
 
 
-def copy_to_readonly_numpy_array_or_list(v, kind=None, force_numeric=False):
+def copy_to_readonly_numpy_array(v, kind=None, force_numeric=False):
     """
     Convert an array-like value into a read-only numpy array
 
@@ -94,7 +94,6 @@ def copy_to_readonly_numpy_array_or_list(v, kind=None, force_numeric=False):
         "i": "int32",
         "f": "float64",
         "O": "object",
-        "U": "U",
     }
 
     # Handle pandas Series and Index objects
@@ -119,12 +118,18 @@ def copy_to_readonly_numpy_array_or_list(v, kind=None, force_numeric=False):
     if not isinstance(v, np.ndarray):
         # v has its own logic on how to convert itself into a numpy array
         if is_numpy_convertable(v):
-            return copy_to_readonly_numpy_array_or_list(
+            return copy_to_readonly_numpy_array(
                 np.array(v), kind=kind, force_numeric=force_numeric
             )
         else:
             # v is not homogenous array
-            return [to_scalar_or_list(e) for e in v]
+            v_list = [to_scalar_or_list(e) for e in v]
+
+            # Lookup dtype for requested kind, if any
+            dtype = kind_default_dtypes.get(first_kind, None)
+
+            # construct new array from list
+            new_v = np.array(v_list, order="C", dtype=dtype)
     elif v.dtype.kind in numeric_kinds:
         # v is a homogenous numeric array
         if kind and v.dtype.kind not in kind:
@@ -135,12 +140,6 @@ def copy_to_readonly_numpy_array_or_list(v, kind=None, force_numeric=False):
         else:
             # Either no kind was requested or requested kind is satisfied
             new_v = np.ascontiguousarray(v.copy())
-    elif v.dtype.kind == "O":
-        if kind:
-            dtype = kind_default_dtypes.get(first_kind, None)
-            return np.array(v, dtype=dtype)
-        else:
-            return v.tolist()
     else:
         # v is a non-numeric homogenous array
         new_v = v.copy()
@@ -155,12 +154,12 @@ def copy_to_readonly_numpy_array_or_list(v, kind=None, force_numeric=False):
     if "U" not in kind:
         # Force non-numeric arrays to have object type
         # --------------------------------------------
-        # Here we make sure that non-numeric arrays become lists
-        # This works around cases like np.array([1, 2, '3']) where
+        # Here we make sure that non-numeric arrays have the object
+        # datatype. This works around cases like np.array([1, 2, '3']) where
         # numpy converts the integers to strings and returns array of dtype
         # '<U21'
         if new_v.dtype.kind not in ["u", "i", "f", "O", "M"]:
-            return v.tolist()
+            new_v = np.array(v, dtype="object")
 
     # Set new array to be read-only
     # -----------------------------
@@ -399,7 +398,7 @@ class DataArrayValidator(BaseValidator):
             # Pass None through
             pass
         elif is_homogeneous_array(v):
-            v = copy_to_readonly_numpy_array_or_list(v)
+            v = copy_to_readonly_numpy_array(v)
         elif is_simple_array(v):
             v = to_scalar_or_list(v)
         else:
@@ -610,7 +609,7 @@ class EnumeratedValidator(BaseValidator):
                 self.raise_invalid_elements(invalid_els[:10])
 
             if is_homogeneous_array(v):
-                v = copy_to_readonly_numpy_array_or_list(v)
+                v = copy_to_readonly_numpy_array(v)
             else:
                 v = to_scalar_or_list(v)
         else:
@@ -766,7 +765,7 @@ class NumberValidator(BaseValidator):
         elif self.array_ok and is_homogeneous_array(v):
             np = get_module("numpy")
             try:
-                v_array = copy_to_readonly_numpy_array_or_list(v, force_numeric=True)
+                v_array = copy_to_readonly_numpy_array(v, force_numeric=True)
             except (ValueError, TypeError, OverflowError):
                 self.raise_invalid_val(v)
 
@@ -893,7 +892,7 @@ class IntegerValidator(BaseValidator):
             pass
         elif self.array_ok and is_homogeneous_array(v):
             np = get_module("numpy")
-            v_array = copy_to_readonly_numpy_array_or_list(
+            v_array = copy_to_readonly_numpy_array(
                 v, kind=("i", "u"), force_numeric=True
             )
 
@@ -1054,7 +1053,25 @@ class StringValidator(BaseValidator):
                 if invalid_els:
                     self.raise_invalid_elements(invalid_els)
 
-            if is_simple_array(v) or is_homogeneous_array(v):
+            if is_homogeneous_array(v):
+                np = get_module("numpy")
+
+                # If not strict, let numpy cast elements to strings
+                v = copy_to_readonly_numpy_array(v, kind="U")
+
+                # Check no_blank
+                if self.no_blank:
+                    invalid_els = v[v == ""][:10].tolist()
+                    if invalid_els:
+                        self.raise_invalid_elements(invalid_els)
+
+                # Check values
+                if self.values:
+                    invalid_inds = np.logical_not(np.isin(v, self.values))
+                    invalid_els = v[invalid_inds][:10].tolist()
+                    if invalid_els:
+                        self.raise_invalid_elements(invalid_els)
+            elif is_simple_array(v):
                 if not self.strict:
                     v = [StringValidator.to_str_or_unicode_or_none(e) for e in v]
 
@@ -1331,12 +1348,8 @@ class ColorValidator(BaseValidator):
             # Pass None through
             pass
         elif self.array_ok and is_homogeneous_array(v):
-            v = copy_to_readonly_numpy_array_or_list(v)
-            if (
-                not isinstance(v, list)
-                and self.numbers_allowed()
-                and v.dtype.kind in ["u", "i", "f"]
-            ):
+            v = copy_to_readonly_numpy_array(v)
+            if self.numbers_allowed() and v.dtype.kind in ["u", "i", "f"]:
                 # Numbers are allowed and we have an array of numbers.
                 # All good
                 pass
@@ -1350,9 +1363,9 @@ class ColorValidator(BaseValidator):
 
                 # ### Check that elements have valid colors types ###
                 elif self.numbers_allowed() or invalid_els:
-                    v = copy_to_readonly_numpy_array_or_list(validated_v, kind="O")
+                    v = copy_to_readonly_numpy_array(validated_v, kind="O")
                 else:
-                    v = copy_to_readonly_numpy_array_or_list(validated_v, kind="U")
+                    v = copy_to_readonly_numpy_array(validated_v, kind="U")
         elif self.array_ok and is_simple_array(v):
             validated_v = [self.validate_coerce(e, should_raise=False) for e in v]
 
@@ -1867,7 +1880,7 @@ class FlaglistValidator(BaseValidator):
                 self.raise_invalid_elements(invalid_els)
 
             if is_homogeneous_array(v):
-                v = copy_to_readonly_numpy_array_or_list(validated_v, kind="U")
+                v = copy_to_readonly_numpy_array(validated_v, kind="U")
             else:
                 v = to_scalar_or_list(v)
         else:
@@ -1915,7 +1928,7 @@ class AnyValidator(BaseValidator):
             # Pass None through
             pass
         elif self.array_ok and is_homogeneous_array(v):
-            v = copy_to_readonly_numpy_array_or_list(v, kind="O")
+            v = copy_to_readonly_numpy_array(v, kind="O")
         elif self.array_ok and is_simple_array(v):
             v = to_scalar_or_list(v)
         return v